Use the Host port when port forwarding is detected
In vagrant and other systems, you often have a port on the host
forwarding to a port on the client box. In terms of ServerUrl, the port
on the host server is the public one that should be used.

Prior to this patch ServerUrl was detecting both the port from the Host
header, as well as SERVER_PORT, and using *both* in generated URLs.

As an example, given:

    HTTP_HOST: localhost:10088
    SERVER_PORT: 10081

ServerUrl would generate the host "localhost:10088:10081", when it
should be generating "localhost:10088". This patch corrects the
situation.