No invokable configuration
It's better to define the services properly up front, and not use the
`invokables` configuration for the plugin manager. One issue that arises
with the `invokables` configuration is that it can override existing
factories and aliases, which may not be what is intended. Defining them
explicitly ensures that incoming configuration at instantiation can
properly override the default services.