Ensure that we get the parent locator
Some tests are injecting the plugin manager as the service locator, when, in
fact, the two instances should be discretely different. This patch ensures the
tests continue to pass by updating `setServiceLocator()` to test if the instance
is a plugin manager. If it is, it pulls the parent context and uses that
instead.