Ensure all builds pass
This patch does a number of things to ensure all builds will pass on
Travis.

First, it removes the `composer.lock`, and modifies the tests to run
against latest and lowest dependencies only, with the following notes:

- When running against lowest, we need to remove zend-router, as
  composer 1.8 incorrectly flags version 3.0.1 incompatible with
  zend-mvc v2 releases, and thus installs zend-mvc 3.0.0 always. I will
  be reporting this to the composer team.

- When running against latest, we need to add
  zend-mvc-plugin-flashmessenger, zend-mvc-i18n, and zend-mvc-console so
  that tests against all helpers will pass.

Through experimentation, I found that running against the lockfile was
problematic due to some dependencies only running on PHP 7.1 and up, as
well as the various mixes in dependencies when running against different
zend-mvc versions. A lowest/latest approach should be sufficient for our
needs until we are able to bump to PHP 7.1+ and drop support for
zend-mvc v2 releases.

Additionally, this patch adds environments for the PHP 7.3 series, to
ensure we are compatible.