Improve type inference for stateful container implementations
Adds templates to the container and abstract placeholder view helper classes and documents the types used by each implementing view helper.

A lot of incorrect docblocks have been fixed, additional magic methods documented and a decent reduction in the baseline.

Several implementations store values as plain objects and were in places (parameter) type hinting on `stdClass`. All references to `stdClass` have been replaced with `object` and `instanceof` conditions replaced with `is_object`. There's no BC break because `object` is more general than `stdClass`. The main purpose here is that it's not possible to express object shapes with `stdClass` which is central to the significant improvements in type inference.

Various public and protected methods have been marked `@internal` so that they can be made private in the next major, and there are a handful of deprecations for unused, un-documented and un-tested public methods.

Signed-off-by: George Steel <george@net-glue.co.uk>