Refactor Escape Helpers
Removes deprecated methods:

- `setEncoding()`
- `getEncoding()`
- `setEscaper()`
- `getEscaper()`
- `setView()`
- `getView()`

Removes inheritance of `AbstractHelper`

Introduces a config provider that allows the service manager to produce an escaper instance configured with a user supplied or default encoding.

Adds a factory for the escape helpers that retrieves the configured escaper and injects that into each helper instance.

Improves tests and verifies uncovered behaviour of escape helpers.

Signed-off-by: George Steel <george@net-glue.co.uk>