Firm up behaviour of HTML Attribute Handling
Boolean values should be handled in a consistent way where `false` attributes are omitted and `true` attributes are encoded in the most compatible way, i.e. `name="name"` which works for XHTML, HTML5 and XML

Array attribute values should be limited to contain stringable/scalar members and should be joined with a predictable separator.

Some well-known attributes should be joined with `','` and some with `' '`. This commit also sets the default separator to a space for unknown attributes.

Attributes are now sorted alphabetically by attribute name to improve output consistency.

Adds a number of tests to cover uncovered behaviour

Signed-off-by: George Steel <george@net-glue.co.uk>