Remove `ViewModel` "Options"
The view model does not need options as part of its contract. The only feasible option might be 'template', but this should be a property of the view model rather than an option.

Users can create their own view models based on `ModelInterface` if options are necessary for some reason.

Signed-off-by: George Steel <george@net-glue.co.uk>