Remove `HelperInterface` requiring all plugins to be `callable`
`HelperInterface` is empty and is purely an affordance for legacy helpers that never implemented an `__invoke` method.

Removing the interface and all affordances means that users will be forced to drop the interface and add at a minimum `public function __invoke(): self;`.

This patch is not necessary and the change could be delayed until v4

Signed-off-by: George Steel <george@net-glue.co.uk>