Prevent ambiguous arguments to `PhpRenderer::render`
`PhpRenderer::render` should be called with `(ViewModel)` or `(string, iterable)`.

When called with `(ViewModel, iterable)`, an exception will now be thrown instead of silently ignoring the `iterable` representing the view variables.

This will be useful signalling to users because in previous versions, the view model and the given variables would have been merged.

Signed-off-by: George Steel <george@net-glue.co.uk>